+++ cmd/transporter/goja_buider_test.go
@@ -11,19 +11,19 @@ import (
 
 func TestNewBuilder(t *testing.T) {
 	a := buildAdaptor("mongodb")(map[string]interface{}{"uri": "mongo://localhost:27017"})
-	source, err := pipeline.NewNode("source", a.name, "test./.*/", a.a, nil)
+	source, err := pipeline.NewNode("source", a.name, DefaultNamespace, a.a, nil)
 	if err != nil {
 		t.Fatalf("unexpected error, %s\n", err)
 	}
 
 	a = buildAdaptor("elasticsearch")(map[string]interface{}{"uri": "http://localhost:9200"})
-	sink, err := pipeline.NewNode("sink", a.name, "test./.*/", a.a, source)
+	sink, err := pipeline.NewNode("sink", a.name, DefaultNamespace, a.a, source)
 	if err != nil {
 		t.Fatalf("unexpected error, %s\n", err)
 	}
 
 	transformer := buildFunction("transformer")(map[string]interface{}{"filename": "pipeline.js"})
-	sink.Transforms = []*pipeline.Transform{&pipeline.Transform{Name: "trans", Fn: transformer, NsFilter: regexp.MustCompile(".*")}}
+	sink.Transforms = []*pipeline.Transform{&pipeline.Transform{Name: "trans", Fn: transformer, NsFilter: regexp.MustCompile(DefaultNamespace)}}
 
 	expected := "Transporter:\n"
 	expected += source.String()
@@ -41,18 +41,18 @@ func TestNewBuilder(t *testing.T) {
 func TestNewBuilderWithEnv(t *testing.T) {
 	os.Setenv("TEST_MONGO_URI", "mongo://localhost:27017")
 	a := buildAdaptor("mongodb")(map[string]interface{}{"uri": "mongo://localhost:27017"})
-	source, err := pipeline.NewNode("source", a.name, "test./.*/", a.a, nil)
+	source, err := pipeline.NewNode("source", a.name, DefaultNamespace, a.a, nil)
 	if err != nil {
 		t.Fatalf("unexpected error, %s\n", err)
 	}
 	a = buildAdaptor("elasticsearch")(map[string]interface{}{"uri": "http://localhost:9200"})
-	sink, err := pipeline.NewNode("sink", a.name, "test./.*/", a.a, source)
+	sink, err := pipeline.NewNode("sink", a.name, DefaultNamespace, a.a, source)
 	if err != nil {
 		t.Fatalf("unexpected error, %s\n", err)
 	}
 
 	transformer := buildFunction("transformer")(map[string]interface{}{"filename": "pipeline.js"})
-	sink.Transforms = []*pipeline.Transform{&pipeline.Transform{Name: "trans", Fn: transformer, NsFilter: regexp.MustCompile(".*")}}
+	sink.Transforms = []*pipeline.Transform{&pipeline.Transform{Name: "trans", Fn: transformer, NsFilter: regexp.MustCompile(DefaultNamespace)}}
 
 	expected := "Transporter:\n"
 	expected += source.String()